+++ kythe/go/services/xrefs/xrefs.go
@@ -849,3 +849,11 @@ func RegisterHTTPHandlers(ctx context.Context, xs Service, mux *http.ServeMux) {
 		}
 	})
 }
+
+// ByName implements the sort.Interface for xpb.Facts
+type ByName []*xpb.Fact
+
+// Implement the sort.Interface
+func (s ByName) Len() int           { return len(s) }
+func (s ByName) Less(i, j int) bool { return s[i].Name < s[j].Name }
+func (s ByName) Swap(i, j int)      { s[i], s[j] = s[j], s[i] }